+++ react/ProductSKUAttributes.js
@@ -1,0 +1,3 @@
+import ProductSKUAttributes from './components/ProductSKUAttributes/Wrapper'
+
+export default ProductSKUAttributes