+++ react/ProductSKUNames.js
@@ -1,3 +1,0 @@
-import ProductSKUNames from './components/ProductSKUNames/Wrapper'
-
-export default ProductSKUNames